Harden pak header/entry validation
Strict load-time validator: PACK signature check, file size captured
via ftell, overflow-safe diroffset+dirlength bounds, per-entry
offset/length bounds, num_entries cap, every fseek/fread/calloc
return checked. In-memory filename buffer widened by one byte and
force-NUL'd so full-width names no longer overread into adjacent
offset/length on strlen/strcmp/printf.

add_file seeks to max(offset+length) across all entries rather than
the directory tail — paks like id's pak0.pak have entries out of
byte order and the old behavior overwrote live data. Input paths
>= 56 bytes rejected before the strcpy that previously overflowed.
add_folder uses bounded snprintf joins, caps recursion at 100, and
closedirs on every exit path.

filesize() widened to int64_t with UINT32_MAX cap; zero-length files
no longer infinite-loop in add_file. extract/copy paths check
malloc/fseek/fread and skip I/O for zero-length entries. Struct
serialization split into per-field writes since the in-memory and
on-disk layouts now diverge.

8 new bats regressions; the non-linear-pak add case fails without
the new max(offset+length) compute.

+++ src/filesystem.h
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stdint.h>
 #include <string.h>
 #include <stdlib.h>
 #include <errno.h>
@@ -7,4 +8,4 @@
 
 int file_exists(const char *dirname);
 int mkdir_r(char *path);
-int filesize (FILE *fd);
+int64_t filesize(FILE *fd);
